+++ wayland-scanner/tests/scanner_assets/test-server-code.rs
@@ -136,6 +136,8 @@ pub mod test_global {
     pub const REQ_DESTROY_SINCE: u32 = 4u32;
     #[doc = r" The minimal object version supporting this request"]
     pub const REQ_REVERSE_LINK_SINCE: u32 = 5u32;
+    #[doc = r" The minimal object version supporting this request"]
+    pub const REQ_NEWID_AND_ALLOW_NULL_SINCE: u32 = 5u32;
     #[doc = r" The minimal object version supporting this event"]
     pub const EVT_MANY_ARGS_EVT_SINCE: u32 = 1u32;
     #[doc = r" The minimal object version supporting this event"]
@@ -176,6 +178,12 @@ pub mod test_global {
         Destroy,
         #[doc = "reverse link a secondary and a tertiary\n\n\n\nOnly available since version 5 of the interface"]
         ReverseLink { sec: Option<super::secondary::Secondary>, ter: super::tertiary::Tertiary },
+        #[doc = "a newid request that also takes allow null arg\n\n\n\nOnly available since version 5 of the interface"]
+        NewidAndAllowNull {
+            quad: New<super::quad::Quad>,
+            sec: Option<super::secondary::Secondary>,
+            ter: super::tertiary::Tertiary,
+        },
     }
     #[derive(Debug)]
     #[non_exhaustive]
@@ -388,6 +396,63 @@ pub mod test_global {
                         Ok((
                             me,
                             Request::ReverseLink {
+                                sec: if sec.is_null() {
+                                    None
+                                } else {
+                                    Some(
+                                        match <super::secondary::Secondary as Resource>::from_id(
+                                            conn,
+                                            sec.clone(),
+                                        ) {
+                                            Ok(p) => p,
+                                            Err(_) => {
+                                                return Err(DispatchError::BadMessage {
+                                                    msg,
+                                                    interface: Self::interface().name,
+                                                })
+                                            }
+                                        },
+                                    )
+                                },
+                                ter: match <super::tertiary::Tertiary as Resource>::from_id(
+                                    conn,
+                                    ter.clone(),
+                                ) {
+                                    Ok(p) => p,
+                                    Err(_) => {
+                                        return Err(DispatchError::BadMessage {
+                                            msg,
+                                            interface: Self::interface().name,
+                                        })
+                                    }
+                                },
+                            },
+                        ))
+                    } else {
+                        Err(DispatchError::BadMessage { msg, interface: Self::interface().name })
+                    }
+                }
+                6u16 => {
+                    if let [Argument::NewId(quad), Argument::Object(sec), Argument::Object(ter)] =
+                        &msg.args[..]
+                    {
+                        Ok((
+                            me,
+                            Request::NewidAndAllowNull {
+                                quad: New::wrap(
+                                    match <super::quad::Quad as Resource>::from_id(
+                                        conn,
+                                        quad.clone(),
+                                    ) {
+                                        Ok(p) => p,
+                                        Err(_) => {
+                                            return Err(DispatchError::BadMessage {
+                                                msg,
+                                                interface: Self::interface().name,
+                                            })
+                                        }
+                                    },
+                                ),
                                 sec: if sec.is_null() {
                                     None
                                 } else {